// src/enums/np-formula-base.ts
export enum NPFormulaBase {
  Premise = "Premise",
  Assumption = "Assumption",
  Shortcut = "Shortcut",
  DI = "Disjunction Introduction",
  DE = "Disjunction Elimination",
  II = "Implication Introduction",
  IE = "Implication Elimination",
  CI = "Conjunction Introduction",
  CE = "Conjunction Elimination",
  EI = "Equivalence Introduction",
  EE = "Equivalence Elimination",
  NE = "Negation Introduction",
  NI = "Negation Elimination",
}